+++ tutorials/viewer_stream/viewer_stream_device.cpp
@@ -237,7 +237,7 @@ void renderTileStandard(int taskIndex,
     RTCRay& ray = rays[N++];
 
     ray.org = Vec3fa(camera.xfm.p);
-    ray.dir = Vec3fa(normalize(x*camera.xfm.l.vx + y*camera.xfm.l.vy + camera.xfm.l.vz));
+    ray.dir = Vec3fa(normalize((float)x*camera.xfm.l.vx + (float)y*camera.xfm.l.vy + camera.xfm.l.vz));
     bool mask = 1; { // invalidates inactive rays
       ray.tnear = mask ? 0.0f         : (float)(pos_inf);
       ray.tfar  = mask ? (float)(inf) : (float)(neg_inf);
